Add schema system and Header property to Manifold
- Integrated MATLAB Zarr support as dependency
- Created bct.manifold.schema package with validation and normalization
- Added schema-driven Header property to bct.Manifold class
- Consolidated ID and Metric into Header structure
- Auto-populate Header with defaults and metadata on construction
- Added setHeader() method for updating metadata
- Manifold.read() now sets Name, Source, CreatedAt, CreatedBy automatically
- Fixed type system: Faces/Edges now uint32, Vertices now double
- Added metadata fields: IndexBase, FaceWinding, NormalConvention, CoordinateSystem, Units
- Updated all dependent functions to use Header.ID and Header.Metric

+++ toolbox/+bct/+manifold/in.m
@@ -64,9 +64,11 @@ if isempty(V) || isempty(F)
         'Object contains empty Vertices or Faces data.');
 end
 
-% Convert to double if necessary (Faces might be int32)
+% Normalize types according to schema
+% Vertices → double (for precision in geometry/operators)
+% Faces → uint32 (semantic correctness and memory efficiency)
 V = double(V);
-F = double(F);
+F = uint32(F);
 
 % Construct Manifold object
 M = bct.Manifold(V, F);
